StaticWebPages

StaticWebPages.academicons

StaticWebPages.color_to_card

StaticWebPages.color_to_label

StaticWebPages.color_to_timeline

StaticWebPages.content

StaticWebPages.info

StaticWebPages.local_info

StaticWebPages.publication_labels

StaticWebPages.user_to_name

StaticWebPages.AbstractItem

StaticWebPages.AbstractSection

StaticWebPages.BackgroundColor

StaticWebPages.Block

StaticWebPages.Block

StaticWebPages.Block

StaticWebPages.Card

StaticWebPages.CardColor

StaticWebPages.ColorLabel

StaticWebPages.Deck

StaticWebPages.Deck

StaticWebPages.Dot

StaticWebPages.Double

StaticWebPages.Double

StaticWebPages.Email

StaticWebPages.FieldSort

StaticWebPages.Git

StaticWebPages.Git

StaticWebPages.GitRepo

StaticWebPages.GitRepo

StaticWebPages.Link

StaticWebPages.Nest

StaticWebPages.Page

StaticWebPages.Parser

StaticWebPages.Publications

StaticWebPages.Publications

StaticWebPages.Section

StaticWebPages.Section

StaticWebPages.TimeLine

StaticWebPages.TimeLine

StaticWebPages.TimeLineColor

StaticWebPages.GitBuilder

StaticWebPages.email

StaticWebPages.export_site

StaticWebPages.head

StaticWebPages.iframe

StaticWebPages.images

StaticWebPages.link

StaticWebPages.nav

StaticWebPages.page

StaticWebPages.paragraphs

StaticWebPages.to_html

StaticWebPages.to_name

StaticWebPages.upload_site

StaticWebPages.academicons — Constant

academicons

A dictionary that tracks the HTML/CSS attributes for icons from font-awesome.

StaticWebPages.color_to_card — Constant

color_to_card

A dictionary to translate the enumeration in CardColor to actual HTML/CSS attributes. A card can be composed of either one or two colors.

StaticWebPages.color_to_label — Constant

color_to_label

A dictionary to translate the enumeration in ColorLabel to actual HTML/CSS labels. An optional black-text color for text is given (it defaults to white-text).

StaticWebPages.color_to_timeline — Constant

color_to_timeline

A dictionary to translate the enumeration in TimeLineColor to actual HTML/CSS attributes.

StaticWebPages.content — Constant

content

An ordered dictionary to store the different web pages.

StaticWebPages.info — Constant

info

A dictionary, with default values, used to customize personal information.

StaticWebPages.local_info — Constant

local_info

A dictionary that stores local information required to build and export the website.

StaticWebPages.publication_labels — Constant

publication_labels

An ordered dictionary to store the attributions of labels to keywords in a Publications.

StaticWebPages.user_to_name — Constant

user_to_name

A dictionary to translate user id (such as GitHub id) to real name. Normal use is that if an entry is found, then the id will be replaced by the appropriate name. Otherwise, the id is kept in use.

When a list of id is too long, priority is given to the id with an associated name.

StaticWebPages.AbstractItem — Type

AbstractItem

An abstract supertype for all items making the content of Sections.

StaticWebPages.AbstractSection — Type

AbstractSection

An abstract supertype for all Sections.

StaticWebPages.BackgroundColor — Type

FieldSort

An enumeration of the different background colors available. It will need to be adjusted for Theme and SubTheme.

bg_none

bg_white

bg_grey

StaticWebPages.Block — Type

Block

An item to construct a sequence of paragraphs with a (possibly empty) sequence of images on the side. Instead of images, an iframe's url can be given.

StaticWebPages.Block — Method

Block(paragraphs, images)

Construct a Block with paragraphs and images.

StaticWebPages.Block — Method

Block(paragraphs, iframe::String)

Construct a Block with paragraphs and an iframe.

StaticWebPages.Card — Type

Card

A Card structure to fill a Deck item.

StaticWebPages.CardColor — Type

CardColor

An enumeration of the different background colors available. It will need to be adjusted for Theme and SubTheme.

card_blue

card_green

card_red

card_orange

cardjuliablue

cardjuliagreen

cardjuliapurple

cardjuliared

StaticWebPages.ColorLabel — Type

ColorLabel

An enumeration of the different colors available for labels.

red

green

yellow

blue

orange

purple

cyan

magenta

lime

pink

teal

lavender

brown

beige

maroon

mint

olive

apricot

navy

grey

white

black

Currently, label colors cycle once the limit has been reached. Creation of cycles with different varieties of colors is welcomed.

StaticWebPages.Deck — Type

Deck

A Deck item, composed of cards.

StaticWebPages.Deck — Method

Deck(color::CardColor, args...)
Deck(args...; color::CardColor=card_blue)

Construct a Deck with the cards in args. The Deck color can be set to: card_blue, card_green, card_red, card_orange, card_julia_blue, card_julia_green, card_julia_purple, card_julia_red.

StaticWebPages.Dot — Type

Dot

A structure for each Dot entry in a TimeLine.

StaticWebPages.Double — Type

Double

A structure for double columns sections.

StaticWebPages.Double — Method

Double(first::Section, second::Section)

Construct a two-columns section.

StaticWebPages.Email — Type

Email

Email inline Item. Can be obfuscated.

StaticWebPages.FieldSort — Type

FieldSort

An enumeration of the different sorting rules for a bibliography item.

unsorted

required

lexicographic

StaticWebPages.Git — Type

Git

A structure to store the data of a git repository.

StaticWebPages.Git — Method

Git(gh_rl::RepoLabels, git_filter)

Create a Git element with the elements found at gh_rl and filtered with git_filter.

StaticWebPages.GitRepo — Type

GitRepo

A structure to store the URLs of git repositories and an associated filter for possible bot filtering.

StaticWebPages.GitRepo — Method

GitRepo(args::RepoLabels...; filter=["github-actions[bot]"])

Stores the URLs of git repositories and an associated filter. By default filters "github-actions[bot]".

StaticWebPages.Link — Type

Link

Structure to handle html links within an <: AbstractItem. Behaviour is similar to the <a> html markup.

StaticWebPages.Nest — Type

Nest

A structure to handle nested items.

StaticWebPages.Page — Type

Page

A structure to store a page information.

StaticWebPages.Parser — Type

Parser

An enumeration of the different parsers for a bibliography item.

bibtex

StaticWebPages.Publications — Type

Publications

A structure that store all the informations required to build and export a bibloigraphy.

StaticWebPages.Publications — Method

Publications(source::String; parser::Parser=bibtex, sort::FieldSort=required)

Import a bibliography from source using parser. Some sorting options can be given through sort. Please check the Bibliography.jl package.

StaticWebPages.Section — Type

Section

A structure for single column sections.

StaticWebPages.Section — Method

Section(; keyargs...)

Create a new single column section.

bgcolor::BackgroundColor = bg_none: set the section background to bg_none (default background), bg_white, or bg_grey.

hide::Bool = false: if set to true, the section will not be built.

items::SectionItems=Nest(): the list of items in the section. Default to an empty Nested item.

title::String="".

title_size=0 : If set to 0, the size will be inherited.

)

StaticWebPages.TimeLine — Type

TimeLine

A structure for a TimeLine item. It is compose of Dots.

StaticWebPages.TimeLine — Method

TimeLine(color::TimeLineColor, args::Dot...)
TimeLine(args::Dot...; color=tl_blue)

Construct a TimeLime with the dots in args. The TimeLine color can be set to: tl_blue, tl_green, tl_red, tl_orange, tl_julia_blue, tl_julia_green, tl_julia_purple, tl_julia_red, tl_julia.

StaticWebPages.TimeLineColor — Type

TimeLineColor

An enumeration of the different TimeLinecolors available. It will need to be adjusted forThemeandSubTheme`.

tl_blue

tl_green

tl_red

tl_orange

tljuliablue

tljuliagreen

tljuliapurple

tljuliared

tl_julia

StaticWebPages.GitBuilder — Method

GitBuilder(r, contributors)

Generate the Git struct according to the repository r. If r does not have all its fields specified (i.e. Nothing), then a default value is used instead.

Return g::Git

StaticWebPages.email — Method

email(address::String; content::String="contact", obfuscated::Bool=true)
email(e::Email)

Construct an email link. Can be obfuscated (default) to avoid bots spams.

StaticWebPages.export_site — Method

export_site(; d::Dict{String,String}=local_info, rm_dir::Bool=false, opt_in::Bool=false)

Export the generated website based on d (default to local_info). The content and site folders need to be valid. If rm_dir is true, the site folder will be deleted before the website is generated.

Users can choose to support StaticWebPages.jl by setting opt_in to true. This will add a small banner in the side navigation menu stating "This website was generated using StaticWebPages.jl" and links to the GitHub repository.

StaticWebPages.head — Method

head(info::Dict{String,String}, page::String)

Generate the head section for a page given the global info.

StaticWebPages.iframe — Method

iframe(url)

Construct an iframe from url.

StaticWebPages.images — Method

images(args::Image...)

Construct an ordered sequence of images for a Block item. An Image should be given as a Pair of strings.

StaticWebPages.link — Method

link(content::String, href::String)
link(l::Link)

Construct an html link within an <: AbstractItem. Behaviour is similar to the <a> html markup.

StaticWebPages.nav — Method

nav(info::Dict{String,String}, content::OrderedDict{String,Any}, page::String, opt_in::Bool)

Generate the navigation menu for a page given its content, and the global info. Please set opt_in to true to promote StaticWebPages.

StaticWebPages.page — Method

page(; keyargs...)

Constructor for Page.

background::BackgroundColor=bg_grey: by default, the first section of a page has a grey background.

hide::Bool=false: an hidden page is not generated. However, it will appear in the navigation menu.

sections::Vector{<:AbstractSection}=Vector{<:AbstractSection}(): list of sections that composes this page.

title::String="": Page's title, can be left empty.

StaticWebPages.paragraphs — Method

paragraphs(args::String...)

Construct an ordered sequence of paragraphs (text entries) for a Block item.

StaticWebPages.to_html — Function

to_html(args...)

Generate the html code associated with each element in the args collection.

StaticWebPages.to_name — Method

to_name(user::String)

Transform a user name as a repository collaborator to name (user). The name is taken from the global user_to_name dictionary.

StaticWebPages.upload_site — Function

upload_site(d::Dict{String,String}=local_info)

Will upload the generated website according to the info in d.