Internals

Internals

StaticWebPages.academicons — Constant

academicons

A dictionary that tracks the HTML/CSS attributes for icons from font-awesome.

StaticWebPages.color_to_card — Constant

color_to_card

A dictionary to translate the enumeration in CardColor to actual HTML/CSS attributes. A card can be composed of either one or two colors.

StaticWebPages.color_to_label — Constant

color_to_label

A dictionary to translate the enumeration in ColorLabel to actual HTML/CSS labels. An optional black-text color for text is given (it defaults to white-text).

StaticWebPages.color_to_timeline — Constant

color_to_timeline

A dictionary to translate the enumeration in TimeLineColor to actual HTML/CSS attributes.

StaticWebPages.content — Constant

content

An ordered dictionary to store the different web pages.

StaticWebPages.info — Constant

info

A dictionary, with default values, used to customize personal information.

StaticWebPages.local_info — Constant

local_info

A dictionary that stores local information required to build and export the website.

StaticWebPages.publication_labels — Constant

publication_labels

An ordered dictionary to store the attributions of labels to keywords in a Publications.

StaticWebPages.user_to_name — Constant

user_to_name

A dictionary to translate user id (such as GitHub id) to real name. Normal use is that if an entry is found, then the id will be replaced by the appropriate name. Otherwise, the id is kept in use.

When a list of id is too long, priority is given to the id with an associated name.

StaticWebPages.AbstractSection — Type

AbstractSection

An abstract supertype for all Sections.

StaticWebPages.BackgroundColor — Type

FieldSort

An enumeration of the different background colors available. It will need to be adjusted for Theme and SubTheme.

bg_none

bg_white

bg_grey

StaticWebPages.ColorLabel — Type

ColorLabel

An enumeration of the different colors available for labels.

red

green

yellow

blue

orange

purple

cyan

magenta

lime

pink

teal

lavender

brown

beige

maroon

mint

olive

apricot

navy

grey

white

black

Currently, label colors cycle once the limit has been reached. Creation of cycles with different varieties of colors is welcomed.

StaticWebPages.Deck — Type

Deck

A Deck item, composed of cards.

StaticWebPages.Deck — Method

Deck(color::CardColor, args...)
Deck(args...; color::CardColor=card_blue)

Construct a Deck with the cards in args. The Deck color can be set to: card_blue, card_green, card_red, card_orange, card_julia_blue, card_julia_green, card_julia_purple, card_julia_red.

StaticWebPages.Dot — Type

Dot

A structure for each Dot entry in a TimeLine.

StaticWebPages.Double — Type

Double

A structure for double columns sections.

StaticWebPages.Double — Method

Double(first::Section, second::Section)

Construct a two-columns section.

StaticWebPages.Email — Type

Email

Email inline Item. Can be obfuscated.

StaticWebPages.FieldSort — Type

FieldSort

An enumeration of the different sorting rules for a bibliography item.

unsorted

required

lexicographic

StaticWebPages.Git — Type

Git

A structure to store the data of a git repository.

StaticWebPages.Git — Method

Git(gh_rl::RepoLabels, git_filter)

Create a Git element with the elements found at gh_rl and filtered with git_filter.

StaticWebPages.Link — Type

Link

Structure to handle html links within an <: AbstractItem. Behaviour is similar to the <a> html markup.

StaticWebPages.Nest — Type

Nest

A structure to handle nested items.

StaticWebPages.Page — Type

Page

A structure to store a page information.

StaticWebPages.Parser — Type

Parser

An enumeration of the different parsers for a bibliography item.

bibtex

StaticWebPages.Publications — Type

Publications

A structure that store all the informations required to build and export a bibloigraphy.

StaticWebPages.Publications — Method

Publications(source::String; parser::Parser=bibtex, sort::FieldSort=required)

Import a bibliography from source using parser. Some sorting options can be given through sort. Please check the Bibliography.jl package.

StaticWebPages.Section — Type

Section

A structure for single column sections.

StaticWebPages.Section — Method

Section(; keyargs...)

Create a new single column section.

bgcolor::BackgroundColor = bg_none: set the section background to bg_none (default background), bg_white, or bg_grey.

hide::Bool = false: if set to true, the section will not be built.

items::SectionItems=Nest(): the list of items in the section. Default to an empty Nested item.

title::String="".

title_size=0 : If set to 0, the size will be inherited.

)

StaticWebPages.GitBuilder — Method

GitBuilder(r, contributors)

Generate the Git struct according to the repository r. If r does not have all its fields specified (i.e. Nothing), then a default value is used instead.

Return g::Git

StaticWebPages.email — Method

email(address::String; content::String="contact", obfuscated::Bool=true)
email(e::Email)

Construct an email link. Can be obfuscated (default) to avoid bots spams.

StaticWebPages.head — Method

head(info::Dict{String,String}, page::String)

Generate the head section for a page given the global info.

StaticWebPages.link — Method

link(content::String, href::String)
link(l::Link)

Construct an html link within an <: AbstractItem. Behaviour is similar to the <a> html markup.

StaticWebPages.nav — Method

nav(info::Dict{String,String}, content::OrderedDict{String,Any}, page::String, opt_in::Bool)

Generate the navigation menu for a page given its content, and the global info. Please set opt_in to true to promote StaticWebPages.

StaticWebPages.page — Method

page(; keyargs...)

Constructor for Page.

background::BackgroundColor=bg_grey: by default, the first section of a page has a grey background.

hide::Bool=false: an hidden page is not generated. However, it will appear in the navigation menu.

sections::Vector{<:AbstractSection}=Vector{<:AbstractSection}(): list of sections that composes this page.

title::String="": Page's title, can be left empty.

StaticWebPages.to_html — Function

to_html(args...)

Generate the html code associated with each element in the args collection.

StaticWebPages.to_name — Method

to_name(user::String)

Transform a user name as a repository collaborator to name (user). The name is taken from the global user_to_name dictionary.