Public

Public

StaticWebPages.AbstractItem — Type

AbstractItem

An abstract supertype for all items making the content of Sections.

StaticWebPages.Block — Type

Block

An item to construct a sequence of paragraphs with a (possibly empty) sequence of images on the side. Instead of images, an iframe's url can be given.

StaticWebPages.Block — Method

Block(paragraphs, images)

Construct a Block with paragraphs and images.

StaticWebPages.Block — Method

Block(paragraphs, iframe::String)

Construct a Block with paragraphs and an iframe.

StaticWebPages.Card — Type

Card

A Card structure to fill a Deck item.

StaticWebPages.CardColor — Type

CardColor

An enumeration of the different background colors available. It will need to be adjusted for Theme and SubTheme.

card_blue

card_green

card_red

card_orange

cardjuliablue

cardjuliagreen

cardjuliapurple

cardjuliared

StaticWebPages.GitRepo — Type

GitRepo

A structure to store the URLs of git repositories and an associated filter for possible bot filtering.

StaticWebPages.GitRepo — Method

GitRepo(args::RepoLabels...; filter=["github-actions[bot]"])

Stores the URLs of git repositories and an associated filter. By default filters "github-actions[bot]".

StaticWebPages.TimeLine — Type

TimeLine

A structure for a TimeLine item. It is compose of Dots.

StaticWebPages.TimeLine — Method

TimeLine(color::TimeLineColor, args::Dot...)
TimeLine(args::Dot...; color=tl_blue)

Construct a TimeLime with the dots in args. The TimeLine color can be set to: tl_blue, tl_green, tl_red, tl_orange, tl_julia_blue, tl_julia_green, tl_julia_purple, tl_julia_red, tl_julia.

StaticWebPages.TimeLineColor — Type

TimeLineColor

An enumeration of the different TimeLinecolors available. It will need to be adjusted forThemeandSubTheme`.

tl_blue

tl_green

tl_red

tl_orange

tljuliablue

tljuliagreen

tljuliapurple

tljuliared

tl_julia

StaticWebPages.export_site — Method

export_site(; d::Dict{String,String}=local_info, rm_dir::Bool=false, opt_in::Bool=false)

Export the generated website based on d (default to local_info). The content and site folders need to be valid. If rm_dir is true, the site folder will be deleted before the website is generated.

Users can choose to support StaticWebPages.jl by setting opt_in to true. This will add a small banner in the side navigation menu stating "This website was generated using StaticWebPages.jl" and links to the GitHub repository.

StaticWebPages.iframe — Method

iframe(url)

Construct an iframe from url.

StaticWebPages.images — Method

images(args::Image...)

Construct an ordered sequence of images for a Block item. An Image should be given as a Pair of strings.

StaticWebPages.paragraphs — Method

paragraphs(args::String...)

Construct an ordered sequence of paragraphs (text entries) for a Block item.

StaticWebPages.upload_site — Function

upload_site(d::Dict{String,String}=local_info)

Will upload the generated website according to the info in d.