StaticWebPages

StaticWebPages.GitBuilder

StaticWebPages.GitBuilder — Method

GitBuilder(r, contributors)

Generate the Git struct according to the repository r. If r does not have all its fields specified (i.e. Nothing), then a default value is used instead.

Return g::Git